JuceDetailModal.tsx 생성
- 기타 스타일 수정
- modal 파일 따로 생성 후 적용
- hover 효과 적용
